Analyse the GDP of different States and generate insights.
The overall goal of this project is to help the CMs focus on areas which will foster economic development for their respective states. Since the most common measure of economic development is the GDP,The task was to analyse the GDP of the various states of India and suggested ways to improve it.
I used Jupyter notebook and anaconda for this project. I have to get the Data for various states from online portals and then collaborate it into one so that further analysis can be performed. After that Data cleaning and inspection was done to drop nulls values and unnecessary Data. Then Data preparation and EDA was performed on the data set. I plotted the various plots for analyze the trends.