House Pricing: Advanced regression Case Study
A US-based housing company named Surprise Housing has decided to enter the Australian market. The company uses data analytics to purchase houses at a price below their actual values and flip them at a higher price.
It is required to model the price of houses with the available independent variables.